Extracted serialization from rtprotocol
