+++ CMakeLists.txt
@@ -12,6 +12,7 @@ add_library(${PROJECT_NAME}
         RTPacket.cpp
         RTProtocol.cpp
         Settings.cpp
+        MarkupSerializer.cpp
 )
 
 target_include_directories(${PROJECT_NAME} PUBLIC
